chore: move function that checks node_modules existence to file-writer.js file
