+++ lib/file-writer.js
@@ -97,6 +97,23 @@ function copyLicenseFiles (dependencyLicenseFiles) {
   });
 }
 
+// Checks if node_modules exists and returns
+// true if it is not empty not considering '.bin'
+// directory.
+const hasNodeModules = (dir) => {
+  const modulesDir = path.join(dir, 'node_modules');
+  if (fs.existsSync(modulesDir)) {
+    const content = fs.readdirSync(modulesDir).filter(e => e !== '.bin');
+    if (content.length > 0) {
+      return true;
+    } else {
+      return false;
+    }
+  } else {
+    return false;
+  }
+};
+
 module.exports = {
   createXml,
   createHtml,
@@ -104,5 +121,6 @@ module.exports = {
   createJson,
   createYaml,
   createLicenseDir,
-  copyLicenseFiles
+  copyLicenseFiles,
+  hasNodeModules
 };
